@@ -1,23 +1,25 @@
 package org.example.model;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.JoinColumn;
+import jakarta.persistence.*;
+import java.math.BigDecimal;
+import java.util.List;
 
 @Entity
+@Table(name = "accounts")
 public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String accountName;
-    private double balance;
-
+    @OneToMany(mappedBy = "account", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Operation> operations;
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
+    private BigDecimal balance;
+    private String name;
+
+    public Account() {
+    }
 
     public Long getId() {
         return id;
@@ -27,20 +29,28 @@ public class Account {
         this.id = id;
     }
 
-    public String getAccountName() {
-        return accountName;
+    public List<Operation> getOperation() {
+        return operations;
     }
 
-    public void setAccountName(String accountName) {
-        this.accountName = accountName;
+    public void setOperation(List<Operation> operations) {
+        this.operations = operations;
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
     }
 
     public User getUser() {
@@ -49,5 +59,16 @@ public class Account {
 
     public void setUser(User user) {
         this.user = user;
+    }
+
+    @Override
+    public String toString() {
+        return "Account{"
+                + "id=" + id
+                + ", operations=" + operations
+                + ", user=" + user
+                + ", balance=" + balance
+                + ", name='" + name + '\''
+                + '}';
     }
 }
