@@ -1,43 +1,45 @@
 package org.example.service;
 
+import org.example.library.Inject;
+import org.example.library.Service;
+import org.example.model.Operation;
+import org.example.model.Report;
+import org.example.model.User;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.Date;
+import java.io.PrintWriter;
+import java.time.LocalDate;
 import java.util.List;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.example.model.Operation;
-import org.example.model.Account;
-import org.springframework.stereotype.Service;
 
 @Service
 public class ReportServiceImpl implements ReportService {
-
-    private static final Logger logger = LoggerFactory.getLogger(ReportServiceImpl.class);
+    @Inject
+    private OperationService operationService;
 
     @Override
-    public List<Operation> getOperationsInPeriod(Account account, Date fromDate, Date toDate) {
-        return null;
+    public Report generateFinancialReport(User user, LocalDate fromDate, LocalDate toDate) {
+        List<Operation> operations = operationService
+                .findOperationsByUserAndPeriod(user, fromDate, toDate);
+        return new Report(user, fromDate, toDate, operations);
     }
 
     @Override
-    public void exportReportToCsv(List<Operation> operations, String filePath) {
-        try (FileWriter writer = new FileWriter(filePath)) {
-            writer.append("OperationType,Amount,Category,Date\n");
-
-            for (Operation operation : operations) {
-                writer.append(String.format("%s,%.2f,%s,%s\n",
-                        operation.getOperationType(),
-                        operation.getAmount(),
-                        operation.getCategory().toString(),
-                        operation.getDate().toString()));
+    public void exportReport(Report report, String filePath) {
+        try (PrintWriter writer = new PrintWriter(new FileWriter(filePath))) {
+            writer.println("ID, Account ID, Name, Category, Amount, Created At");
+            for (Operation operation : report.getOperations()) {
+                writer.println(
+                        operation.getId() + ", " +
+                                operation.getAccount().getId() + ", " +
+                                operation.getName() + ", " +
+                                operation.getCategory() + ", " +
+                                operation.getAmount() + ", " +
+                                operation.getCreatedAt());
             }
 
-            logger.info("CSV файл успешно создан!");
+            System.out.println("Report exported successfully to: " + filePath);
         } catch (IOException e) {
-            logger.error("Ошибка при создании CSV файла: {}", e.getMessage());
+            System.out.println("Error exporting report to CSV: " + e.getMessage());
         }
     }
-
 }
