@@ -1,13 +1,15 @@
-package org.example.service;
+package org.example.dao;
 
 import org.example.model.Account;
 import org.example.model.Operation;
-
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
-public interface AccountService {
+public interface AccountDao {
+    List<Operation> findOperationsInPeriod
+            (Account account, LocalDate fromDate, LocalDate toDate);
+
     Account add(Account account);
 
     void update(Account account);
@@ -15,7 +17,4 @@ public interface AccountService {
     void delete(Account account);
 
     Optional<Account> findById(Long id);
-
-    List<Operation> findOperationsInPeriod(Account account,
-                                           LocalDate fromDate, LocalDate toDate);
 }
