@@ -1,37 +1,51 @@
 package org.example.service;
 
+import org.example.dao.OperationDao;
+import org.example.library.Inject;
+import org.example.library.Service;
+import org.example.model.Account;
 import org.example.model.Operation;
-import org.example.repository.OperationRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.example.model.User;
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class OperationServiceImpl implements OperationService {
+    @Inject
+    private OperationDao operationDao;
 
-    private final OperationRepository operationRepository;
-
-    @Autowired
-    public OperationServiceImpl(OperationRepository operationRepository) {
-        this.operationRepository = operationRepository;
+    @Override
+    public Operation add(Operation operation) {
+        if (operation.getAmount() <= 0 || operation.getCategory() == null) {
+            throw new RuntimeException("Error: The amount of the operation must be positive "
+                    + "and operation must have a category");
+        }
+        return operationDao.add(operation);
     }
 
     @Override
-    public void createOperation(Operation operation) {
-        operationRepository.save(operation);
+    public void update(Operation operation) {
+        operationDao.update(operation);
     }
 
     @Override
-    public Operation getOperationById(Long operationId) {
-        return operationRepository.findById(operationId).orElse(null);
+    public void delete(Operation operation) {
+        operationDao.delete(operation);
     }
 
     @Override
-    public void updateOperation(Operation operation) {
-        operationRepository.save(operation);
+    public Optional<Operation> findById(Long id) {
+        return operationDao.findById(id);
     }
 
     @Override
-    public void deleteOperation(Long operationId) {
-        operationRepository.deleteById(operationId);
+    public List<Operation> findOperationsByAccount(Account account) {
+        return operationDao.findOperationsByAccount(account);
+    }
+
+    @Override
+    public List<Operation> findOperationsByUserAndPeriod(User user, LocalDate fromDate, LocalDate toDate) {
+        return operationDao.findOperationsByUserAndPeriod(user, fromDate, toDate);
     }
 }
