@@ -1,50 +1,33 @@
 package org.example.service;
 
+import org.example.dao.UserDao;
+import org.example.library.Inject;
+import org.example.library.Service;
 import org.example.model.User;
-import org.example.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import javax.persistence.EntityNotFoundException;
+import java.util.Optional;
+
 @Service
 public class UserServiceImpl implements UserService {
+    @Inject
+    private UserDao userDao;
 
-    private final UserRepository userRepository;
-
-    @Autowired
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
+    @Override
+    public User add(User user) {
+        return userDao.add(user);
     }
 
     @Override
-    public void createUser(User user) {
-        if (user.getName() == null || user.getEmail() == null || userRepository.existsByEmail(user.getEmail())) {
-            throw new IllegalArgumentException("Invalid user data");
-        }
-
-        userRepository.save(user);
+    public void update(User user) {
+        userDao.update(user);
     }
 
     @Override
-    public User getUserById(final Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(new java.util.function.Supplier<EntityNotFoundException>() {
-                    @Override
-                    public EntityNotFoundException get() {
-                        return new EntityNotFoundException("User not found with id: " + userId);
-                    }
-                });
-    }
-
-
-
-
-    @Override
-    public void updateUser(User user) {
-        userRepository.save(user);
+    public void delete(User user) {
+        userDao.delete(user);
     }
 
     @Override
-    public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+    public Optional<User> findById(Long id) {
+        return userDao.findById(id);
     }
 }
